chore: rename to user_metadata; add typeshare + doc comments
