+++ .vscode/settings.json
@@ -10,6 +10,7 @@
     "pubkey",
     "Solana",
     "tryorama",
+    "typeshare",
     "workdir",
     "zome",
     "zomes"
