+++ Cargo.toml
@@ -8,9 +8,9 @@ opt-level = "z"
 resolver = "2"
 members = [
   "crates/hdk_utils",
-  "crates/metadata/types",
-  "crates/metadata/handlers",
-  "crates/metadata/validation",
+  "crates/user_metadata/types",
+  "crates/user_metadata/handlers",
+  "crates/user_metadata/validation",
   "crates/holoom_dna_tests",
   "crates/holoom_types",
   "crates/jaq_wrapper",
@@ -39,18 +39,19 @@ holochain = { version = "=0.4.0-dev.12", default-features = false, features = [
   "test_utils",
 ] }
 holochain_keystore = "=0.4.0-dev.12"
+typeshare = "1"
 
 [workspace.dependencies.hdk_utils]
 path = "crates/hdk_utils"
 
-[workspace.dependencies.metadata_types]
-path = "crates/metadata/types"
+[workspace.dependencies.user_metadata_types]
+path = "crates/user_metadata/types"
 
-[workspace.dependencies.metadata_validation]
-path = "crates/metadata/validation"
+[workspace.dependencies.user_metadata_validation]
+path = "crates/user_metadata/validation"
 
 [workspace.dependencies.user_metadata_handlers]
-path = "crates/metadata/handlers"
+path = "crates/user_metadata/handlers"
 
 [workspace.dependencies.holoom_types]
 path = "crates/holoom_types"
