+++ crates/user_metadata/handlers/Cargo.toml
@@ -10,7 +10,6 @@ name = "user_metadata_handlers"
 [dependencies]
 serde = { workspace = true }
 hdk = { workspace = true }
-metadata_types = { workspace = true }
+user_metadata_types = { workspace = true }
 bincode = { workspace = true }
 hdk_utils = { workspace = true }
-ts-rs = "9"
