+++ crates/username_registry_coordinator/Cargo.toml
@@ -15,9 +15,10 @@ username_registry_integrity = { workspace = true }
 holoom_types = { workspace = true }
 username_registry_validation = { workspace = true }
 username_registry_utils = { workspace = true }
-metadata_types = { workspace = true }
+user_metadata_types = { workspace = true }
 user_metadata_handlers = { workspace = true }
 jaq_wrapper = { workspace = true }
 indexmap = "2.2.6"
 serde_json = { workspace = true }
 hex = { workspace = true }
+typeshare = { workspace = true }
