+++ crates/username_registry_integrity/Cargo.toml
@@ -12,5 +12,5 @@ hdi = { workspace = true }
 serde = { workspace = true }
 holoom_types = { workspace = true }
 username_registry_validation = { workspace = true }
-metadata_types = { workspace = true }
-metadata_validation = { workspace = true }
+user_metadata_types = { workspace = true }
+user_metadata_validation = { workspace = true }
